inventory-ai/requirements.txt
numpy>=1.26,<3.0
pandas>=2.0,<3.0
matplotlib>=3.8
scikit-learn>=1.3
prophet==1.3.0
cmdstanpy==1.3.0
tqdm>=4.66
holidays==0.95
SQLAlchemy>=2.0
PyMySQL>=1.1
python-dotenv>=1.0
Flask>=2.3
